+++ src/team/types.d.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright 2022 steadybit GmbH. All rights reserved.
+ */
+
+export interface TeamSummary {
+  teams: Team[];
+}
+
+export interface Team {
+  key: string;
+  name: string;
+  allowedAreas: string[]
+}